Url logic added
Here i tried to add the pattern to check for urls and so far it's working even if it's not working well for all the urls.

Example can be ` https://www.youtube.com/watch?v=0Z7RjfeQaNk&pp=0gcJCdgAo7VqN5tD` where am not getting all the url as it's and when all emails are incorporated in one text it tends not to extract them all.