Phone number logic added
Here i adde the functionality for the Phone numbers and unfortunately am not getting the country codes at all and to make it valid you have to sepate the numbers with the comma or simply by dot.

And also the numbers have to nine without including the country codes.

And as addition to that, i added `\b` at the beginning and end of the patterns so that we can be able to allocate where to validate without getting stuck in sentences.